refactored the classes and removed the Rebellious Hierarchy Design smell from the project
