+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/parser/address/AddressTraceReader.java
@@ -18,7 +18,7 @@ package com.github.benmanes.caffeine.cache.simulator.parser.address;
 import java.util.stream.LongStream;
 
 import com.github.benmanes.caffeine.cache.simulator.parser.TextTraceReader;
-import com.github.benmanes.caffeine.cache.simulator.parser.TraceReader.KeyOnlyTraceReader;
+import com.github.benmanes.caffeine.cache.simulator.parser.KeyOnlyTraceReader;
 
 /**
  * A reader for the trace files of application address instructions, provided by
