+++ simulator/src/main/java/com/github/benmanes/caffeine/cache/simulator/parser/gradle/GradleTraceReader.java
@@ -19,7 +19,7 @@ import java.math.BigInteger;
 import java.util.stream.LongStream;
 
 import com.github.benmanes.caffeine.cache.simulator.parser.TextTraceReader;
-import com.github.benmanes.caffeine.cache.simulator.parser.TraceReader.KeyOnlyTraceReader;
+import com.github.benmanes.caffeine.cache.simulator.parser.KeyOnlyTraceReader;
 
 /**
  * A reader for the Gradle Build Cache trace files provided by the Gradle team.
